Fix worlds as well

+++ patches/tModLoader/Terraria/ModLoader/IO/WorldIO.cs
@@ -488,13 +488,14 @@ namespace Terraria.ModLoader.IO
 		internal static void EraseWorld(string path, bool cloudSave) {
 			path = Path.ChangeExtension(path, ".twld");
 			if (!cloudSave) {
-#if WINDOWS
-				FileOperationAPIWrapper.MoveToRecycleBin(path);
-				FileOperationAPIWrapper.MoveToRecycleBin(path + ".bak");
-#else
-				File.Delete(path);
-				File.Delete(path + ".bak");
-#endif
+				if (OperatingSystem.IsWindows()) {
+					FileOperationAPIWrapper.MoveToRecycleBin(path);
+					FileOperationAPIWrapper.MoveToRecycleBin(path + ".bak");
+				}
+				else {
+					File.Delete(path);
+					File.Delete(path + ".bak");
+				}
 			}
 			else if (SocialAPI.Cloud != null) {
 				SocialAPI.Cloud.Delete(path);
